requirements.txt
aioice==0.10.2
aiortc==1.14.0
annotated-types==0.7.0
av==16.1.0
cffi==2.0.0
cryptography==46.0.5
cyberwave==0.3.41
dnspython==2.8.0
google-crc32c==1.8.0
ifaddr==0.2.0
numpy==2.4.3
opencv-python==4.13.0.92
paho-mqtt==2.1.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
